feat(flat): output_format field + with_output_format + QueryView.flat_output
Adds Dataset.output_format (Literal["ragged","flat"], default "ragged") as
the last defaulted field in the frozen dataclass, a with_output_format()
method that validates and replaces the field, and threads flat_output=bool
into QueryView so later tasks can consume it. No runtime behavior changes.

Moves snap_dataset fixture from test_flat_getitem_snapshot.py into
tests/dataset/conftest.py so it is shared by both the snapshot suite and
the new test_output_format.py.
