+++ tests/dataset/conftest.py
@@ -1,0 +1,54 @@
+"""Shared fixtures for tests/dataset/."""
+
+from __future__ import annotations
+
+import pytest
+import pyBigWig
+
+import genvarloader as gvl
+
+SEQLEN = 20
+
+
+@pytest.fixture(scope="session")
+def snap_dataset(source_bed, vcf_dir, reference, tmp_path_factory):
+    """Phased VCF dataset with a "5ss" BigWig track, opened with a reference.
+
+    Mirrors the ``base_ds`` fixture in ``tests/dataset/test_with_methods.py``.
+    Opened with default settings (output_length="ragged", sequence_type="haplotypes",
+    jitter=0, max_jitter=2, deterministic=True, rc_neg=True).
+    """
+    from genoray import VCF
+
+    tmp_dir = tmp_path_factory.mktemp("snap_ds")
+    out = tmp_dir / "snap.gvl"
+
+    vcf_samples = ["s0", "s1", "s2"]
+    # Header lengths are generous upper bounds for the regions in source.bed.
+    contig_sizes = [("chr1", 2_000_000), ("chr2", 2_000_000)]
+    bw_paths: dict[str, str] = {}
+    for i, sample in enumerate(vcf_samples):
+        bw_path = tmp_dir / f"{sample}.bw"
+        with pyBigWig.open(str(bw_path), "w") as bw:
+            bw.addHeader(contig_sizes, maxZooms=0)
+            # One short interval per contig region in source.bed; values differ
+            # per sample. Mirrors base_ds in tests/dataset/test_with_methods.py.
+            value = float(i + 1)
+            bw.addEntries(
+                ["chr1", "chr1", "chr2", "chr2"],
+                [499_990, 1_010_686, 17_320, 1_234_560],
+                ends=[500_030, 1_010_706, 17_340, 1_234_580],
+                values=[value, value, value, value],
+            )
+        bw_paths[sample] = str(bw_path)
+
+    bigwigs = gvl.BigWigs("5ss", bw_paths)
+    vcf = VCF(vcf_dir / "filtered_source.vcf.gz")
+    gvl.write(
+        path=out,
+        bed=source_bed,
+        variants=vcf,
+        tracks=bigwigs,
+        max_jitter=2,
+    )
+    return gvl.Dataset.open(out, reference=reference)
